requirements.txt
selenium>=4.0.0
pytest>=7.0.0
pytest-html>=3.0.0
pytest-xdist>=2.0.0
webdriver-manager>=3.8.0 
tenacity~=9.0.0
Faker~=33.1.0
requests~=2.32.3
psutil~=7.0.0
allure-python-commons~=2.14.3
screenrecord~=1.1.3